% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/art18.R
\docType{data}
\name{art18}
\alias{art18}
\title{Arthurs and Timmers 2018 consonance ratings}
\format{
A data frame with 12 rows and 6 variables:
\describe{
\item{id}{Chord names}
\item{name}{Unique identifier of chords}
\item{rating}{Mean ratings}
\item{pi_chord}{pitch of the chord tones in midi}
\item{chord_size}{Number of pitches in the chord, numerosity}
\item{dataset}{Name of the dataset}
...
}
}
\source{
\url{https://doi.org/10.1177/0305735617720834}
}
\usage{
art18
}
\description{
A dataset containing ratings of consonance.
}
\keyword{datasets}
