% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jl12b.R
\docType{data}
\name{jl12b}
\alias{jl12b}
\title{Johnson-Laird et al 2012, Exp. 2 consonance ratings}
\format{
A data frame with 48 rows and 5 variables:
\describe{
\item{id}{Unique identifier of chords}
\item{rating}{Mean ratings}
\item{roughness_model}{Roughness model output}
\item{tonal_model}{Tonal model output}
\item{pi_chord}{pitch of the chord tones in midi}
...
}
}
\source{
\url{https://doi.org/10.1525/mp.2012.30.1.19}
}
\usage{
jl12b
}
\description{
A dataset containing ratings of consonance.
}
\keyword{datasets}
