% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lah20b.R
\docType{data}
\name{lah20b}
\alias{lah20b}
\title{Lahdelma and Eerola 2020 Exp. 2 consonance ratings}
\format{
A data frame with 72 rows and 10 variables:
\describe{
\item{id}{Unique identifier of chords}
\item{rating}{Mean ratings}
\item{rating_sd}{standard deviation of the ratings}
\item{rating_se}{standard error of the ratings}
\item{pi_chord}{pitch of the chord tones in midi}
\item{chord_size}{Number of pitches in the chord, numerosity}
\item{dataset}{Name of the dataset}
\item{dissonance_level}{Level of dissonance from Bowling}
\item{name}{Name of the chord}
\item{chord_nro}{Number of chord in the experiment}
...
}
}
\source{
\url{https://doi.org/10.1038/s41598-020-65615-8}
}
\usage{
lah20b
}
\description{
A dataset containing ratings of consonance.
}
\keyword{datasets}
