% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop19.R
\docType{data}
\name{pop19}
\alias{pop19}
\title{Popescu et al 2019 consonance ratings}
\format{
A data frame with 80 rows and 7 variables:
\describe{
\item{id}{Chord id}
\item{genre}{Musical genre representing the origin of the chord}
\item{rating}{Mean rating}
\item{rating_sd}{SD of the rating}
\item{rating_se}{SE of the rating}
\item{pi_chord}{pitch of the chord tones in midi}
\item{chord_size}{Number of pitches in the chord, numerosity}
\item{dataset}{Name of the dataset}
...
}
}
\source{
\url{https://doi.org/10.1038/s41598-018-35873-8}
}
\usage{
pop19
}
\description{
A dataset containing ratings of consonance.
}
\keyword{datasets}
