% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sch03.R
\docType{data}
\name{sch03}
\alias{sch03}
\title{Schwartz et al., 2003 consonance ratings}
\format{
A data frame with 12 rows and 4 variables:
\describe{
\item{id}{Chord names}
\item{pi_chord_type}{pitch-class of the chord tones}
\item{chord_size}{Number of pitches in the chord, numerosity}
\item{rating}{Mean ratings}
...
}
}
\source{
\url{https://doi.org/10.1523/JNEUROSCI.23-18-07160.2003}
}
\usage{
sch03
}
\description{
A dataset containing ratings of consonance.
}
\keyword{datasets}
